requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pandas>=2.0.0
pdfplumber>=0.9.0
python-multipart>=0.0.6
openpyxl>=3.1.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
requests>=2.31.0